# app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dtos.ThemeResponse:
    properties:
      problems:
        items:
          $ref: '#/definitions/models.Problem'
        type: array
      text:
        type: string
    type: object
...
  dtos.ValidationErrorResponse:
    properties:
      error:
        additionalProperties:
          type: string
        type: object
    type: object
...
  dtos.VerifyAnswerRequest:
    properties:
      answer:
        type: string
      language:
        type: string
      problem:
        type: string
    required:
    - answer
    - language
    - problem
    type: object
  models.Problem:
    properties:
      id:
        type: integer
      is_theory:
        type: boolean
      languages:
        items:
          type: string
        type: array
      question:
        type: string
      theme:
        $ref: '#/definitions/models.Theme'
      theme_id:
        type: integer
    type: object
  models.Theme:
    properties:
      id:
        type: integer
      scores:
        type: integer
      title:
        type: string
      topic:
        $ref: '#/definitions/models.Topic'
      topic_id:
        type: integer
    type: object
  models.Topic:
    properties:
      id:
        type: integer
      scores:
        type: integer
      themes:
        items:
          $ref: '#/definitions/models.Theme'
        type: array
      title:
        type: string
    type: object
...
  services.PromptProblemResponse:
    properties:
      message:
        type: string
      ok:
        type: boolean
    type: object
host: https://educgpt.ru
info:
  contact: {}
  title: Educ-GPT API
  version: "1.0"
paths:
...
  /roadmap:
    get:
      consumes:
      - application/json
      description: Returns a list of topics for the current user
      parameters:
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of topics
          schema:
            items:
              $ref: '#/definitions/models.Topic'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get topics
      tags:
      - roadmap
  /roadmap/{topic_id}:
    get:
      consumes:
      - application/json
      description: Returns a list of themes for a specific topic, sorted by user progress
        and AI recommendations
      parameters:
      - description: Topic ID
        in: path
        name: topic_id
        required: true
        type: integer
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of themes
          schema:
            items:
              $ref: '#/definitions/models.Theme'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "409":
          description: AI request error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get themes
      tags:
      - roadmap
  /roadmap/{topic_id}/{theme_id}:
    get:
      consumes:
      - application/json
      description: Returns detailed information about a specific theme, including
        problems and AI-generated content
      parameters:
      - description: Topic ID
        in: path
        name: topic_id
        required: true
        type: integer
      - description: Theme ID
        in: path
        name: theme_id
        required: true
        type: integer
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Theme details
          schema:
            $ref: '#/definitions/dtos.ThemeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Theme or topic not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "409":
          description: AI request error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get theme details
      tags:
      - roadmap
  /roadmap/info/theme/{theme_id}:
    get:
      consumes:
      - application/json
      description: Returns info of theme without authorization
      produces:
      - application/json
      responses:
        "200":
          description: Theme info
          schema:
            $ref: '#/definitions/models.Theme'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get theme info
      tags:
      - roadmap
  /roadmap/info/topic/{topic_id}:
    get:
      consumes:
      - application/json
      description: Returns info of topic without authorization
      produces:
      - application/json
      responses:
        "200":
          description: Topic info
          schema:
            $ref: '#/definitions/models.Topic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get topic info
      tags:
      - roadmap
  /roadmap/problems/{topic_id}/{theme_id}:
    get:
      consumes:
      - application/json
      description: Returns a list of problems for a specific theme, generated by AI
      parameters:
      - description: Topic ID
        in: path
        name: topic_id
        required: true
        type: integer
      - description: Theme ID
        in: path
        name: theme_id
        required: true
        type: integer
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of problems
          schema:
            items:
              $ref: '#/definitions/models.Problem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Theme or topic not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "409":
          description: AI request error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get problems
      tags:
      - roadmap
  /roadmap/resolve:
    post:
      consumes:
      - application/json
      description: VerifyAnswer and get verification status by AI
      parameters:
      - description: Bearer <JWT token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Answer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.VerifyAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Answer verification result
          schema:
            $ref: '#/definitions/services.PromptProblemResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dtos.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "409":
          description: AI request error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Verify answer
      tags:
      - roadmap
...
swagger: "2.0"